+++ harmony/hvigorfile.ts
@@ -1,0 +1,7 @@
+// Botler 鸿蒙端工程级 Hvigor 入口
+import { appTasks } from '@ohos/hvigor-ohos-plugin';
+
+export default {
+  system: appTasks, /* 内置插件，不可修改 */
+  plugins: []       /* 自定义插件扩展位 */
+}